Division linking automation
Adjustment to the division detection algorithm to automate the detection threshold. Now based on an automated loop that aims to maximise permissiveness while still preventing closed loops in the lineage trees. Also fixes an issue that was preventing the use of fluorescence as a feature in the division detection module.